refactor: Moved remove_entrance to waypoint
